myproject/requirements.txt
Django>=3.2,<4.0
djangorestframework
djangorestframework-simplejwt
django-rest-passwordreset
gunicorn
django-cors-headers
django-ratelimit
dj-database-url
psycopg2-binary
pandas
numpy
numpy_financial
openpyxl
python-dotenv
requests
openai
langchain
langchain_experimental